Fix generatepot-rst for api-quick-start
The guide has no common directory, therefore we have no common
translation. Handle this  in our script.

+++ tools/generatepot-rst.sh
@@ -93,8 +93,10 @@ EOF
     rm -f ${TOPDIR}common-rst/source/locale/dummy.po
 else
     # common-rst is translated as part of openstack-manuals, do not
-    # include the file in the combined tree.
-    rm ${DIRECTORY}/source/locale/common.pot
+    # include the file in the combined tree if it exists.
+    if [ -f ${DIRECTORY}/source/locale/common.pot ] ; then
+        rm ${DIRECTORY}/source/locale/common.pot
+    fi
 fi
 
 # Take care of deleting all temporary files so that
